loading SDE values as floats.

+++ model/sde/SDE-Raw/src/main/java/fr/guiguilechat/jcelechat/model/sde/load/bsd/EdgmAttributeTypes.java
@@ -69,7 +69,7 @@ public class EdgmAttributeTypes {
 	public int attributeID;
 	public String attributeName;
 	public int categoryID;
-	public double defaultValue;
+	public float defaultValue;
 	public String description;
 	public boolean highIsGood;
 	public boolean published;
